+++ src/import-paths/index.ts
@@ -1,0 +1,15 @@
+import * as javascript from './javascript';
+import * as typescript from './typescript';
+import * as css from './css';
+import * as scss from './scss';
+import * as html from './html';
+import * as markdown from './markdown';
+
+export {
+  javascript,
+  typescript,
+  css,
+  scss,
+  html,
+  markdown
+};