+++ src/import-texts/index.ts
@@ -1,6 +1,3 @@
-/* 
-  Import text barrel files
-  */
 export * from './html-markdown';
 export * from './css-scss';
 export * from './javascript-typescript';