+++ src/model/index.ts
@@ -1,6 +1,3 @@
-/* 
-  Model barrel files
-  */
 export * from './types';
 export * from './enums';
 export * from './interfaces';