+++ src/model/types.ts
@@ -1,21 +1,21 @@
 /* 
   Supported file extension types
   */
-export type FileExtensions = '.ts' | '.tsx' | '.js' | '.jsx' 
+export type FileExtension = '.ts' | '.tsx' | '.js' | '.jsx' 
   | '.css' | '.scss'
   | '.html' | '.md';
 
 /* 
   Supported image file extension types
   */
-export type ImageFileExtensions = '.gif' | '.jpeg' | '.jpg' | '.png' | '.webp';
+export type ImageFileExtension = '.gif' | '.jpeg' | '.jpg' | '.png' | '.webp';
 
 /* 
   Supported image and file extension types
   */
-export type SupportedFileExtensions = FileExtensions | ImageFileExtensions;
+export type SupportedFileExtension = FileExtension | ImageFileExtension;
 
 /* 
   Get Import text option types
   */
-export type ImportTextOptions = 'script' | 'stylesheet' | 'markdown' | 'image';
+export type ImportType = 'script' | 'stylesheet' | 'markdown' | 'image';