+++ src/modules/index.ts
@@ -1,7 +1,5 @@
-/* 
-  Utils barrel files
-  */
 export * from './file-extension';
+export * from './import-type';
 export * from './notify';
 export * from './relative-path';
 export * from './import-text';